% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop.R
\name{crop}
\alias{crop}
\title{crop}
\usage{
crop(data)
}
\arguments{
\item{data}{data.frame to be cropped}
}
\value{
cropped data.frame
}
\description{
crop removes all columns and rows that are completely missing data. Note this is a cleaner, simpler version of dropAllNA().
}
\examples{
## Example needs to be added
}

