% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dichot.R
\name{dichot}
\alias{dichot}
\title{dichot}
\usage{
dichot(x, group1, group2)
}
\arguments{
\item{x}{input data}

\item{group1}{vector of levels to be set to 0}

\item{group2}{vector of levels to be set to 1}
}
\value{
output vector of 0, 1, NA
}
\description{
dichot will take a variable and dichotomize it into 0, 1 (and NA for any unspecified levels)
}
\examples{
#fakeData <- data.frame(x = sample(c(1, 2, 3, 4, 5), 20, replace = T))
#dichot(x = fakeData$x, group1 = c(1, 2), group2 = c(3, 4))
}

