% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropAllNA.R
\name{dropAllNA}
\alias{dropAllNA}
\title{dropAllNA}
\usage{
dropAllNA(data, axis = c("all", "col", "row"))
}
\arguments{
\item{data}{data.frame containing data}

\item{axis}{which axis to remove from: 'row' removes all rows where all columns have missing data. 'col' removes all columns where all rows are NA. 'all' does both.}
}
\value{
data frame with the rows/columns missing
}
\description{
dropAllNA  will drop all columns, rows, or both, where all entries in the data.frame are NA
}
\examples{
#df <- data.frame(a = c(1, 2, NA, 4),
#                 b = c(5, 6, NA, 7),
#                 c = c(NA, NA, NA, NA),
#                 d = c(8, 9, NA, 10))
#dropAllNA(df) #default is 'all'
#dropAllNA(df, 'all') #remove rows and cols with missing
#dropAllNA(df, 'col') #remove cols with missing
#dropAllNA(df, 'row') #remove rows with missing
}

