% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyCode.R
\name{dummyCode}
\alias{dummyCode}
\title{dummyCode}
\usage{
dummyCode(x, convertNA = TRUE)
}
\arguments{
\item{x}{input data}

\item{convertNA}{TRUE/FALSE: Should NA be converted to a unique level}
}
\value{
A data.frame with dummy coded variables
}
\description{
dummyCode will create a dummy coded set of variables
}
\examples{
#x <- sample(c(1,2,3,4,5), 20, replace = T)
#dummyCode(x = x)
}

