% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical_join.R
\name{logical_join}
\alias{logical_join}
\title{logical_join}
\usage{
logical_join(data1, data2, by = NULL)
}
\arguments{
\item{data1}{data.frame to join as reference dataset}

\item{data2}{data.frame to look for potential matches}

\item{by}{variable or variables to use for joining data1 and data2}
}
\value{
data1 is returned with a boolian variable 'inData2' appended.
}
\description{
logical_join provides booliean variable appended onto data1, indicating which rows have matching observations in data2
}
\examples{
#df1 <- data.frame(id = 1:10, x = rnorm(10))
#df2 <- data.frame(id = sample(c(1:10), 5, replace = F), y = rnorm(10))
#logical_join(df1, df2)
#df1 \%>\% logical_join(df2)
#df1 \%>\% logical_join(df2, by = 'id')
}

