% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(x, method = 2)
}
\arguments{
\item{x}{A numeric vector}

\item{method}{If method=1, standard normalize method used. If method=2, percentile ranks are provided.}
}
\value{
A numeric vector
}
\description{
normalize transforms a numeric vector to percentile ranks
}
\examples{
#x <- seq(1:10)
#x2 = runif(100, 1, 44)
#normalize(x, 1)
#normalize(x, 2)
#normalize(x2, 1)
#normalize(x2, 2)
}

