% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reFactor.R
\name{reFactor}
\alias{reFactor}
\title{reFactor}
\usage{
reFactor(x, order_index = NULL, order_levels = NULL)
}
\arguments{
\item{x}{Vector to be refactored}

\item{order_index}{Index of the current factor levels in which to refactor with}

\item{order_levels}{Actual Vector levels to use to refactor with}
}
\value{
Refactored vector
}
\description{
reFactor will refactor a categorical variable such that the order of the levels have changed (e.g change what the reference group will be)
}
\examples{
#v <- sample(letters[1:3], 10, replace = T)
#factor(v)
#Reverse order
#reFactor(v)
#refactor based on original factor index
#reFactor(v, order_index = c(2, 1, 3))
#refactor based on actual levels
#reFactor(v, order_levels = c("b", "a", "c"))
}

