% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmOutliers.R
\name{rmOutliers}
\alias{rmOutliers}
\title{rmOutliers}
\usage{
rmOutliers(x, sdCut = 3, method = "remove")
}
\arguments{
\item{x}{Vector of continuous values}

\item{sdCut}{Number of standard deviations to use as a cut-point for finding outliers}

\item{method}{Method for outlier handling. 'remove' sets outliers to NA. 'limits' will set outliers to be equal to the mean +/- the standard deviation set. 'average' sets outliers to be equal to the average.}
}
\value{
A numeric vector with no outliers
}
\description{
rmOutliers converts outleirs (as defined as being some x number of standard deviations from the mean) to either NA, mean +/- x standard deviations, or the mean
}
\examples{
#example_vector <- c(100, rnorm(19, 5, 2))
#rmOutliers(example_vector, 3, method = "remove")
#rmOutliers(example_vector, 3, method = "limits")
#rmOutliers(example_vector, 3, method = "average")
}

