% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_lvcf.R
\name{row_lvcf}
\alias{row_lvcf}
\title{row_lvcf}
\usage{
row_lvcf(data, cols)
}
\arguments{
\item{data}{data.frame containing data}

\item{cols}{numbers to carry forward (in the order to carry forward)}
}
\value{
data.frame with last values carried forward
}
\description{
row_lvcf
}
\examples{
# Needs an example
}

