% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_.R
\name{t_}
\alias{t_}
\title{t_}
\usage{
t_(x)
}
\arguments{
\item{x}{matrix, vector, or data.frame or other class usable by t().}
}
\value{
a transposed object
}
\description{
t_ is an expansion of the t() function, in which, if x is a data.frame(), it is transposed, and the first column of the output data.frame is the column names from the input data.frame.
}
\examples{
#example_df <- data.frame(x1 = rnorm(100),
#                        x2 = rnorm(100),
#                        x3 = rnorm(100),
#                         x4 = rnorm(100),
#                         x5 = rnorm(100))
#t(rnorm(100))
#t_(rnorm(100))
#t(example_df)
#t_(example_df)
}

