% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{trim}
\usage{
trim(x, side = "both")
}
\arguments{
\item{x}{string or vector of strings}

\item{side}{Which side should be trimmed: 'both', 'left', or 'right'}
}
\value{
string or vector of strings
}
\description{
trim removes whitespace from the left and right side. Depricated with R v3.0.0 (trimws())
}
\examples{
#x = "bill "
#x2 = c("bill ", " bill", " bill ")
#trim(x)
#trim(x2)
#trim(x2, "left")
#trim(x2, "right")
}

